Enhance chart visibility and update Submit button loading state in CoordinatesForm
This commit introduces updates to enhance the visibility of charts by adjusting grid lines and ticks on the X-axis within the ChartVisual component. Specifically, an additional CSS style has been added to the _highcharts.scss file to define stroke-width for the X-axis grid line, and the tick length on the X-axis has been customised for better visibility.

Furthermore, the loading state of the Submit button in the CoordinatesForm component has been improved. The spinner icon that appears when the form is being submitted